requirements.txt
streamlit
torch
torchvision
numpy
joblib
Pillow
opencv-python
matplotlib
scikit-learn
plotly